Add more testing states, started developing way to add radius to points
Added a state to allow fast loading which only loads data from 1st of July 2020. I have also created another file to test converting a point in a matrix to having a radius. I also changed the method of showing the data from a heatmap to a 2D histogram as it is is more useful for what I am trying to do