Added Sydney data, added comments, added line graph
Added Sydney substation coordinates so that we can finally view Sydney data in this. Added code comments to explain and make the program more readable. Added a line graph to show change in fractal dimension throughout the day. redid the interface to make more sense with different configurations, add nan checks to values